AUT-834: Put microlient in separate module
